+++ packages/convergence/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+- `.append()` method to combine convergences together
+
 - "module" entry point to support native consumption of @bigtest/mocha
   as es module